+++ src/processing/blossoms/install/apt/apt_methods.cpp
@@ -24,7 +24,7 @@
 
 #include <libKitsunemimiCommon/common_methods/string_methods.h>
 #include <libKitsunemimiCommon/common_items/data_items.h>
-#include <processing/process_execution.h>
+#include <processing/common/process_execution.h>
 
 namespace SakuraTree
 {
