+++ src/processing/blossoms/special/sakura_copy_subtree_blossom.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "sakura_copy_subtree_blossom.h"
-#include <processing/process_execution.h>
+#include <processing/common/process_execution.h>
 #include <sakura_root.h>
 
 namespace SakuraTree
