+++ src/processing/blossoms/ssh/ssh_cmd_create_file_blossom.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "ssh_cmd_create_file_blossom.h"
-#include <processing/process_execution.h>
+#include <processing/common/process_execution.h>
 
 namespace SakuraTree
 {
