+++ src/processing/blossoms/ssh/ssh_scp_blossom.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "ssh_scp_blossom.h"
-#include <processing/process_execution.h>
+#include <processing/common/process_execution.h>
 
 namespace SakuraTree
 {
