+++ src/processing/common/functions.h
