+++ tests/processing/blossoms/files/common_files/file_rename_blossom_test.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "file_rename_blossom_test.h"
-#include <processing/process_execution.h>
+#include <processing/common/process_execution.h>
 #include <items/sakura_items.h>
 #include <processing/blossoms/files/common_files/file_rename_blossom.h>
