+++ tests/processing/blossoms/files/file_methods_test.cpp
@@ -22,7 +22,7 @@
 
 #include "file_methods_test.h"
 #include <processing/blossoms/files/file_methods.h>
-#include <processing/process_execution.h>
+#include <processing/common/process_execution.h>
 
 namespace SakuraTree
 {
