+++ tests/processing/blossoms/files/text_files/text_write_blossom_test.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "text_write_blossom_test.h"
-#include <processing/process_execution.h>
+#include <processing/common/process_execution.h>
 #include <items/sakura_items.h>
 #include <processing/blossoms/files/text_files/text_write_blossom.h>
 #include <libKitsunemimiPersistence/files/text_file.h>
