+++ tests/processing/blossoms/special/cmd_blossom_test.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "cmd_blossom_test.h"
-#include <processing/process_execution.h>
+#include <processing/common/process_execution.h>
 #include <items/sakura_items.h>
 #include <processing/blossoms/special/cmd_blossom.h>
 #include <processing/blossoms/files/file_methods.h>
